+++ timesheet/src/main/java/org/catrobat/jira/timesheet/rest/json/JsonTimesheet.java
@@ -32,7 +32,7 @@ public final class JsonTimesheet {
     @XmlElement
     private String reason;
     @XmlElement
-    private int ects;
+    private double ects;
     @XmlElement
     private String latestEntryDate;
     @XmlElement
@@ -52,7 +52,7 @@ public final class JsonTimesheet {
     @XmlElement
     private boolean isMTSheet;
 
-    public JsonTimesheet(int timesheetID, String lectures, String reason, int ects, String latestEntryDate, int targetHourPractice,
+    public JsonTimesheet(int timesheetID, String lectures, String reason, double ects, String latestEntryDate, int targetHourPractice,
                          int targetHourTheory, int targetHours, int targetHoursCompleted, int targetHoursRemoved, boolean isActive,
                          boolean isEnabled, boolean isMTSheet) {
         this.timesheetID = timesheetID;
@@ -97,11 +97,11 @@ public final class JsonTimesheet {
         this.reason = reason;
     }
 
-    public int getEcts() {
+    public double getEcts() {
         return ects;
     }
 
-    public void setEcts(int ects) {
+    public void setEcts(double ects) {
         this.ects = ects;
     }
 
@@ -195,7 +195,7 @@ public final class JsonTimesheet {
         int result = timesheetID;
         result = 31 * result + lectures.hashCode();
         result = 31 * result + reason.hashCode();
-        result = 31 * result + ects;
+        result = 31 * result + (int)ects;
         //result = 31 * result + latestEntryDate.hashCode();
         result = 31 * result + targetHourPractice;
         result = 31 * result + targetHourTheory;
